[step2] test. 리펙토링에 따른 테스트 변경 및 통계 테스트 분리
